refactor : profile 컴포넌트 공통 컴포넌트 파일로 이동
