+++ src/pages/AssignActivityPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Profile from '../components/profile/Profile';
+import Profile from '../components/common/profile/Profile';
 
 const AssignActivityPage = () => (
   <div>
